# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Тестовое задание",
        "description": "Разработчик: Абдуазимов Бунёд.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/company-balance/{id}": {
            "get": {
                "tags": [
                    "Company"
                ],
                "summary": "Получить баланс компании",
                "operationId": "ac0f6fee1ea8ca15fdbd5bbf06fce52c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор компании",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный запрос"
                    },
                    "401": {
                        "description": "Неавторизованный доступ"
                    },
                    "404": {
                        "description": "Компания не найдена"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/orders/{id}/status": {
            "put": {
                "tags": [
                    "Order"
                ],
                "summary": "Обновить статус заказа",
                "description": "Обновляет статус указанного заказа.",
                "operationId": "updateOrderStatus ",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID заказа",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Статус заказа",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "водитель на месте",
                                "исполняется"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное обновление статуса заказа",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Статус заказа успешно обновлен"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка при обновлении статуса заказа",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Некорректный ID заказа или статус"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/orders-complate/{id}": {
            "put": {
                "tags": [
                    "Order"
                ],
                "summary": "Завершить заказ",
                "description": "Завершает заказ и обновляет балансы водителя и компании.",
                "operationId": "completeOrder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID заказа",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Заказ успешно завершен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Заказ успешно завершен"
                                        },
                                        "driver_balance": {
                                            "type": "number",
                                            "format": "double",
                                            "example": "150.50"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка при завершении заказа",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Некорректный заказ или водитель не выбран"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}